requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
qdrant-client
rank-bm25
pdfplumber
pymupdf4llm
pymupdf4llm[ocr,layout]
pymupdf
google-generativeai
python-multipart